@@ -13,42 +13,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.2",
-    "@angular/common": "^5.0.2",
-    "@angular/compiler": "^5.0.2",
-    "@angular/core": "^5.0.2",
-    "@angular/forms": "^5.0.2",
-    "@angular/http": "^5.0.2",
-    "@angular/platform-browser": "^5.0.2",
-    "@angular/platform-browser-dynamic": "^5.0.2",
-    "@angular/router": "^5.0.2",
+    "@angular/animations": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
     "core-js": "^2.5.1",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.18",
-    "express": "4.16.2",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
+    "zone.js": "^0.8.26",
+    "express": "4.16.3",
     "body-parser": "1.18.2"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "^5.0.2",
-    "@angular/language-service": "^5.0.2",
+    "@angular/cli": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/language-service": "^6.0.0",
     "@types/googlemaps": "3.30.0",
-    "@types/jasmine": "~2.8.2",
+    "@types/jasmine": "~2.8.7",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.0.53",
-    "angular-playground": "^2.3.0",
-    "codelyzer": "~4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@types/node": "~10.0.3",
+    "angular-playground": "^4.0.1",
+    "codelyzer": "~4.3.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.2.0",
+    "protractor": "~5.3.0",
     "ts-node": "~3.3.0",
-    "tslint": "~5.8.0",
-    "typescript": "~2.4.2"
+    "tslint": "~5.10.0",
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.0"
   }
 }
