@@ -3,7 +3,7 @@
 import 'rxjs/add/observable/of';
 import { Injectable } from '@angular/core';
 import { PreloadingStrategy, Route } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
+import { Observable, of } from 'rxjs';
 
 @Injectable()
 export class PreloadModulesStrategy implements PreloadingStrategy {
@@ -13,7 +13,7 @@ export class PreloadModulesStrategy implements PreloadingStrategy {
       console.log('Preloaded: ' + route.path);
       return load();
     } else {
-      return Observable.of(null);
+      return of(null);
     }
   }
 
